move Localizer.java back to util folder
