+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/Localizer.java
@@ -1,12 +1,10 @@
-package org.firstinspires.ftc.teamcode.subassemblies.autonomous.localizers;
+package org.firstinspires.ftc.teamcode.util;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.util.Pose;
-import org.firstinspires.ftc.teamcode.util.Subassembly;
 
 @Config
 public abstract class Localizer extends Subassembly {
